+++ PXCTestKit/Extension/FBSimulatorControl+PXCTestKit.swift
@@ -12,7 +12,7 @@ import Foundation
 extension FBSimulatorControl {
 
     static func withContext(_ context: ControlContext) throws -> FBSimulatorControl {
-        let logFileHandle = try context.outputManager.createLogFile()
+        let logFileHandle = try context.outputManager.createNewSimulatorLogFile()
         return try FBSimulatorControl.withConfiguration(
             FBSimulatorControlConfiguration(deviceSetPath: context.deviceSet.path, options: context.simulatorManagementOptions),
             logger: FBControlCoreLogger.aslLoggerWriting(toFileDescriptor: logFileHandle.fileDescriptor, withDebugLogging: false)